Add files via upload
use the cards as A12345678910JQK a12345678910jqk with suit cards as diamond=d or D ,spades =S or s ,club =C or c ,Hearts =H or h. use the both upper case as well lower case letters in the poket hands and board cards. this is designed for calculating the Tvbet and Vbet Poker live oods .